+++ parallel_implementation/src/include/char_freq.h
@@ -19,7 +19,7 @@ typedef struct CharFreqDictionary {
 	CharFreq *charFreqs; 
 } CharFreqDictionary;
 
-extern void getCharFreqsFromText(CharFreqDictionary* dict, char text[], int len, int pid);
+extern void getCharFreqsFromText(CharFreqDictionary* dict, char text[], long len, int pid);
 
 extern void sortCharFreqs(CharFreqDictionary* res);
 extern void appendToCharFreqs(CharFreqDictionary* dict, char character, int pos);
